+++ concurrency-practice/src/main/java/me/staek/lock/reentrantlock/basic/_03_ReentrantState.java
@@ -1,4 +1,4 @@
-package me.staek.lock.reentrantlock;
+package me.staek.lock.reentrantlock.basic;
 
 import java.util.concurrent.locks.Lock;
 import java.util.concurrent.locks.ReentrantLock;
